refactor(error): Global Error 코드 추가 및 리펙토링
