+++ src/main/java/org/sejong/sulgamewiki/common/exception/GlobalException.java
@@ -1,0 +1,8 @@
+package org.sejong.sulgamewiki.common.exception;
+
+public class GlobalException extends CustomException{
+
+  public GlobalException(ErrorCode errorCode) {
+    super(errorCode, "global");
+  }
+}
