+++ src/main/java/org/sejong/sulgamewiki/common/exception/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package org.sejong.sulgamewiki.common.exception;
 
+import org.sejong.sulgamewiki.common.exception.dto.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
